Add files via upload
Background

The topic of predicting product recommendations using machine learning revolves around leveraging advanced algorithms and data analysis techniques to forecast and suggest products or services that are likely to be of interest to individual customers. By analyzing historical purchase data, user preferences, and other relevant information, machine learning models can be trained to make personalized recommendations, enhancing the customer experience and driving business growth. It is important to investigate the various strategies introduced to achieve a more efficient way of predicting product recommendation in order to better understand which techniques are to be used for the best possible outcome. The primary aim of forecasting is to better understand the patterns that impacts markets and consumer needs. This study utilizes the state-of-the-art methods used to predict product recommendation in the fashion industry that could be implemented in the real world. The methods implemented to forecast, the models and algorithms used within the methods, the datasets used, and also the trade-offs for each system and model used are described in this analysis. In particular, this analysis will further explore Machine Learning models that are widely applied to support the forecasting efforts carried out in the fashion industry. The scope of this analysis is for customers within the age of 20 to 35.